fix: (habit-alert) duplicate query service error
